+++ test/close_to_list_test.dart
@@ -14,12 +14,12 @@ void expectToFail(
 }
 
 void main() {
-  group('closeTo()', () {
+  group('closeToList()', () {
     test('Iterable<int>', () {
       final i1 = [101, 102];
       final i2 = [111, 103];
       final delta = 2;
-      expectToFail(i1, isCloseTo(i2, delta),
+      expectToFail(i1, closeToList(i2, delta),
           failMessage: 'Expected: a numerical iterable differing from '
               '[111, 103] by less than <$delta>\n'
               '  Actual: [101, 102]\n'
@@ -30,23 +30,23 @@ void main() {
     test('Iterable<double>', () {
       final d1 = [101, 102];
       final d2 = [104.0, 102.0];
-      expect(d1, isCloseTo(d2, 4));
+      expect(d1, closeToList(d2, 4));
     });
     test('Iterable<num>', () {
       final n1 = <num>[101.0, 102.0];
       final n2 = <num>[105, 105];
-      expect(n2, isCloseTo(n1, 5));
+      expect(n2, closeToList(n1, 5));
     });
     test('Empty Iterable', () {
-      expect(<int>[], isCloseTo(<num>[], 5));
+      expect(<int>[], closeToList(<num>[], 5));
     });
     test('Infinity', () {
-      expect([double.infinity], isCloseTo([0], double.infinity));
+      expect([double.infinity], closeToList([0], double.infinity));
     });
     test('NaN', () {
       expectToFail(
         [99],
-        isCloseTo([99], double.nan),
+        closeToList([99], double.nan),
         failMessage: 'Expected: a numerical iterable differing from '
             '[99] by less than <NaN>\n'
             '  Actual: [99]\n'